+++ examples/drupal_loadtest.rs
@@ -82,10 +82,10 @@ fn main() {
 
 /// View the front page.
 async fn drupal_loadtest_front_page(client: &mut GooseClient) {
-    let response = client.get("/").await;
+    let mut response = client.get("/").await;
 
     // Grab some static assets from the front page.
-    match response {
+    match response.response {
         Ok(r) => match r.text().await {
             Ok(t) => {
                 let re = Regex::new(r#"src="(.*?)""#).unwrap();
@@ -97,12 +97,12 @@ async fn drupal_loadtest_front_page(client: &mut GooseClient) {
             }
             Err(e) => {
                 eprintln!("failed to parse front page: {}", e);
-                client.set_failure(&GooseMethod::GET, "/");
+                client.set_failure(&mut response.request);
             }
         },
         Err(e) => {
             eprintln!("unexpected error when loading front page: {}", e);
-            client.set_failure(&GooseMethod::GET, "/");
+            client.set_failure(&mut response.request);
         }
     }
 }
@@ -121,8 +121,8 @@ async fn drupal_loadtest_profile_page(client: &mut GooseClient) {
 
 /// Log in.
 async fn drupal_loadtest_login(client: &mut GooseClient) {
-    let response = client.get("/user").await;
-    match response {
+    let mut response = client.get("/user").await;
+    match response.response {
         Ok(r) => {
             match r.text().await {
                 Ok(html) => {
@@ -131,7 +131,7 @@ async fn drupal_loadtest_login(client: &mut GooseClient) {
                         Some(f) => f,
                         None => {
                             eprintln!("no form_build_id on page: /user page");
-                            client.set_failure(&GooseMethod::GET, "/user");
+                            client.set_failure(&mut response.request);
                             return;
                         }
                     };
@@ -152,7 +152,7 @@ async fn drupal_loadtest_login(client: &mut GooseClient) {
                 }
                 Err(e) => {
                     eprintln!("unexpected error when loading /user page: {}", e);
-                    client.set_failure(&GooseMethod::GET, "/user");
+                    client.set_failure(&mut response.request);
                 }
             }
         }
@@ -166,8 +166,8 @@ async fn drupal_loadtest_post_comment(client: &mut GooseClient) {
     let nid: i32 = rand::thread_rng().gen_range(1, 10_000);
     let node_path = format!("node/{}", &nid);
     let comment_path = format!("/comment/reply/{}", &nid);
-    let response = client.get(&node_path).await;
-    match response {
+    let mut response = client.get(&node_path).await;
+    match response.response {
         Ok(r) => {
             match r.text().await {
                 Ok(html) => {
@@ -177,7 +177,7 @@ async fn drupal_loadtest_post_comment(client: &mut GooseClient) {
                         Some(f) => f,
                         None => {
                             eprintln!("no form_build_id found on {}", &node_path);
-                            client.set_failure(&GooseMethod::GET, &node_path);
+                            client.set_failure(&mut response.request);
                             return;
                         }
                     };
@@ -187,7 +187,7 @@ async fn drupal_loadtest_post_comment(client: &mut GooseClient) {
                         Some(f) => f,
                         None => {
                             eprintln!("no form_token found on {}", &node_path);
-                            client.set_failure(&GooseMethod::GET, &node_path);
+                            client.set_failure(&mut response.request);
                             return;
                         }
                     };
@@ -197,7 +197,7 @@ async fn drupal_loadtest_post_comment(client: &mut GooseClient) {
                         Some(f) => f,
                         None => {
                             eprintln!("no form_id found on {}", &node_path);
-                            client.set_failure(&GooseMethod::GET, &node_path);
+                            client.set_failure(&mut response.request);
                             return;
                         }
                     };
@@ -214,8 +214,8 @@ async fn drupal_loadtest_post_comment(client: &mut GooseClient) {
                         ("op", "Save"),
                     ];
                     let request_builder = client.goose_post(&comment_path);
-                    let response = client.goose_send(request_builder.form(&params)).await;
-                    match response {
+                    let mut response = client.goose_send(request_builder.form(&params)).await;
+                    match response.response {
                         Ok(r) => match r.text().await {
                             Ok(html) => {
                                 if !html.contains(&comment_body) {
@@ -223,7 +223,7 @@ async fn drupal_loadtest_post_comment(client: &mut GooseClient) {
                                         "no comment showed up after posting to {}",
                                         &comment_path
                                     );
-                                    client.set_failure(&GooseMethod::POST, &comment_path);
+                                    client.set_failure(&mut response.request);
                                 }
                             }
                             Err(e) => {
@@ -231,7 +231,7 @@ async fn drupal_loadtest_post_comment(client: &mut GooseClient) {
                                     "unexpected error when posting to {}: {}",
                                     &comment_path, e
                                 );
-                                client.set_failure(&GooseMethod::POST, &comment_path);
+                                client.set_failure(&mut response.request);
                             }
                         },
                         // Goose will catch this error.
@@ -240,7 +240,7 @@ async fn drupal_loadtest_post_comment(client: &mut GooseClient) {
                 }
                 Err(e) => {
                     eprintln!("unexpected error when loading {} page: {}", &node_path, e);
-                    client.set_failure(&GooseMethod::GET, &node_path);
+                    client.set_failure(&mut response.request);
                 }
             }
         }